+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_datetime_picker/flutter_datetime_picker.dart';
 import 'package:japanese_prefecture_picker/japanese_prefecture_picker.dart';
 
 void main() {
