+++ lib/japanese_prefecture_picker.dart
@@ -1,4 +1,4 @@
 library japanese_prefecture_picker;
 
-export 'src/japanese_prefecture_picker.dart';
-export 'src/widgets/japanese_prefecture_picker_header.dart';
+export 'src/japanese_address_picker.dart';
+export 'src/widgets/japanese_address_picker_header.dart';
